pyproject.toml
[tool.poetry]
name = "InnoHack Hackathon"
version = "0.1.2-dev2"
description = "InnoHack Hackathon Backend Repository with PostgreSQL(SQLalchemy) FastAPI Firebase Auth"
authors = ["ATISHAY JAIN <atishayj.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
fastapi = "^0.109.0"
firebase-admin = "^6.3.0"
pydantic = "^2.5.3"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
pytz = "^2023.3.post1"
httpx = "^0.26.0"
psycopg2-binary = "^2.9.9"
azure-cognitiveservices-vision-computervision = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
isort = "^5.13.2"
black = "^23.12.1"
bumpversion = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
